@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-ruby '2.0.0'
+ruby '2.2.0'
 
 gem 'capybara'
-gem 'selenium-webdriver'
+gem 'capybara-webkit'
